dom/plugins/test/mochitest/test_refresh_navigator_plugins.html

<!DOCTYPE html>
<!-- bug 820708 -->
<html>
  <head>
    <meta><charset="utf-8"/>
    <title>Test Refreshing navigator.plugins (bug 820708)</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="plugin-utils.js"></script>
    <link rel="stylesheet" type="text/css"
          href="/tests/SimpleTest/test.css">
  </head>
  <body>
    <p id="display"></p>
    <script class="testbody" type="application/javascript">
      "use strict";

      SimpleTest.waitForExplicitFinish();
      setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);

      ok("Test Plug-in" in navigator.plugins, "testplugin should be present");
      ok("application/x-test" in navigator.mimeTypes, "testplugin MIME should be present");

      setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_DISABLED);
      ok(!("Test Plug-in" in navigator.plugins), "testplugin should not be present");
      ok(!("application/x-test" in navigator.mimeTypes), "testplugin MIME should not be present");

      setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);
      ok("Test Plug-in" in navigator.plugins, "testplugin should be present again");
      ok("application/x-test" in navigator.mimeTypes, "testplugin MIME should be present again");
      SimpleTest.finish();
    </script>
  </body>
</html>
